Super fix

+++ rout/deletefollow.js
@@ -7,7 +7,7 @@ app.post('/deletefollow', function(req, res) {
 })
 
 function deletefollow(req, res) {
-  sequelize.query(`DELETE FROM followers WHERE following = '${req.body.id}'`)
+  sequelize.query(`DELETE FROM followers WHERE following = '${req.body.id}' and follower = '${req.body.userId}'`)
   .then((followers) => {
     res.json("Success!");
   })
